+++ R/plotPairs.R
@@ -32,11 +32,13 @@
 #'   opacity of points in the plot.
 #' @param colorByCorrelation Logical scalar, indicating whether the correlation
 #'   panels should be colored according to the correlation value.
+#' @param addIdentityLine Logical scalar, indicating whether the identity line 
+#'   should be added (only used if \code{pointsType = "points"}).
 #'
 #' @importFrom GGally eval_data_col ggpairs wrap
 #' @importFrom ggplot2 ggplot annotate theme_void ylim stat_density2d
 #'   scale_fill_continuous geom_point theme_bw theme element_blank aes
-#'   geom_histogram scale_x_continuous scale_y_continuous
+#'   geom_histogram scale_x_continuous scale_y_continuous geom_abline
 #' @importFrom stats cor
 #' @importFrom SummarizedExperiment assayNames assay
 #' @importFrom grDevices hcl.colors rgb colorRamp
@@ -46,7 +48,7 @@ plotPairs <- function(se, selAssay = "counts", doLog = TRUE, pseudocount = 1,
                       corMethod = "pearson", histBreaks = 40,
                       pointsType = "points", corSizeMult = 5,
                       corSizeAdd = 2, pointSize = 0.1, pointAlpha = 0.3,
-                      colorByCorrelation = TRUE) {
+                      colorByCorrelation = TRUE, addIdentityLine = FALSE) {
 
     .assertVector(x = se, type = "SummarizedExperiment")
     .assertScalar(x = selAssay, type = "character",
@@ -63,6 +65,7 @@ plotPairs <- function(se, selAssay = "counts", doLog = TRUE, pseudocount = 1,
     .assertScalar(x = pointSize, type = "numeric", rngExcl = c(0, Inf))
     .assertScalar(x = pointAlpha, type = "numeric", rngExcl = c(0, Inf))
     .assertScalar(x = colorByCorrelation, type = "logical")
+    .assertScalar(x = addIdentityLine, type = "logical")
 
     ## ----------------------------------------------------------------------- ##
     ## Define shared theme elements
@@ -131,12 +134,22 @@ plotPairs <- function(se, selAssay = "counts", doLog = TRUE, pseudocount = 1,
     }
 
     ## Define function to create scatter plot (for use with ggpairs)
-    plotpoints <- function(data, mapping, ...) {
-        ggplot2::ggplot(data = data, mapping = mapping) +
-            ggplot2::geom_point(alpha = pointAlpha, size = pointSize) +
-            ggtheme
+    if (addIdentityLine) {
+        plotpoints <- function(data, mapping, ...) {
+            ggplot2::ggplot(data = data, mapping = mapping) +
+                ggplot2::geom_abline(slope = 1, intercept = 0, 
+                                     linetype = "dashed", color = "grey") + 
+                ggplot2::geom_point(alpha = pointAlpha, size = pointSize) +
+                ggtheme
+        }
+    } else {
+        plotpoints <- function(data, mapping, ...) {
+            ggplot2::ggplot(data = data, mapping = mapping) +
+                ggplot2::geom_point(alpha = pointAlpha, size = pointSize) +
+                ggtheme
+        }
     }
-
+    
     ## Define the function to use for the plots
     if (pointsType == "smoothscatter") {
         lower <- list(continuous = smoothscat)